refactor: convert "utils" into separated "helpers" (#33)
