+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [v2.4.4](https://github.com/henriquehbr/svelte-typewriter/compare/v2.4.3...v2.4.4)
+## [Unreleased](https://github.com/henriquehbr/svelte-typewriter/compare/v2.4.4...HEAD)
+
+### Commits
+
+- refactor: convert "utils" into separated "helpers" (#33) [`3851067`](https://github.com/henriquehbr/svelte-typewriter/commit/3851067124b9031babdd531a97208e6a26144c5e)
+
+## [v2.4.4](https://github.com/henriquehbr/svelte-typewriter/compare/v2.4.3...v2.4.4) - 2020-12-15
 
 ### Fixed
