+++ lib/esm/Typewriter.mjs
@@ -1,1 +1,1 @@
-import{SvelteComponent as e,init as t,safe_not_equal as o,element as n,append as l,create_slot as a,set_style as s,attr as r,toggle_class as c,insert as i,action_destroyer as d,update_slot as p,transition_in as m,transition_out as u,detach as y}from"svelte/internal";function f(){const e=function(){throw new Error("Function called outside component initialization")}();return(t,o)=>{e.$$.callbacks[t]}}const v=e=>new Promise((t=>setTimeout(t,e))),b=(e,t)=>Math.floor(Math.random()*(t-e)+e),g=e=>1===e.childNodes.length&&3===e.childNodes[0].nodeType,h=async e=>v(e[b(0,e.length)]||e),w=e=>e.childNodes.forEach((e=>e.remove())),x=(e,t)=>{const o=document.createElement(t);return o.textContent=e,o},$=e=>{if(g(e)){const t=e.textContent,o=x(e.textContent,"p");return e.textContent="",e.appendChild(o),[{currentNode:o,text:t}]}return[...e.children].map((e=>{const t=e.innerHTML.replaceAll("&amp;","&");return{currentNode:e,text:t}}))},T=async(e,t)=>{const o=e.innerHTML.replaceAll("&amp;","&");for(let n=o.length-1;n>=0;n--){">"===o[n]&&(n=o.lastIndexOf("<",n)),e.innerHTML=o.slice(0,n),await h(t.interval)}},L=async({currentNode:e,text:t},o)=>{e.textContent="",e.classList.add("typing");for(let n=0;n<=t.length;n++){"<"===t[n]&&(n=t.indexOf(">",n)+1),e.innerHTML=t.slice(0,n);const l=t.replaceAll("&","&amp;");e.innerHTML===l&&(o.dispatch("done"),await v("number"==typeof o.loop?o.loop:1500),await T(e,o)),await h(o.interval)}null!==e.nextSibling&&1==e.classList.length?e.removeAttribute("class"):e.classList.remove("typing")};var C=Object.freeze({__proto__:null,sleep:v,rng:b,hasSingleTextNode:g,typingInterval:h,cleanChildNodes:w,createElement:x,getElements:$,loopTypewriterEffect:L});const N=async(e,t)=>{const{default:o}=t.loop&&await import("./loop.mjs")||t.loopRandom&&await import("./loopRandom.mjs")||t.cascade&&await import("./cascade.mjs")||t.scramble&&await import("./scramble.mjs")||await import("./default.mjs"),n=$(e);if(t.delay>0){const{sleep:o}=await Promise.resolve().then((function(){return C}));await o(t.delay),e.classList.remove("delay")}o({node:e,elements:n},t)};function E(e){let t,o,l,f,v;const b=e[9].default,g=a(b,e,e[8],null);return{c(){t=n("div"),g&&g.c(),s(t,"--cursor-color","string"==typeof e[0]?e[0]:"black"),r(t,"class","svelte-8gtb9p"),c(t,"cursor",e[0]),c(t,"delay",e[1].delay>0)},m(n,a){i(n,t,a),g&&g.m(t,null),l=!0,f||(v=d(o=N.call(null,t,e[1])),f=!0)},p(e,[o]){g&&g.p&&256&o&&p(g,b,e,e[8],o,null,null),(!l||1&o)&&s(t,"--cursor-color","string"==typeof e[0]?e[0]:"black"),1&o&&c(t,"cursor",e[0]),2&o&&c(t,"delay",e[1].delay>0)},i(e){l||(m(g,e),l=!0)},o(e){u(g,e),l=!1},d(e){e&&y(t),g&&g.d(e),f=!1,v()}}}function M(e,t,o){let{$$slots:n={},$$scope:l}=t,{interval:a=30}=t,{cascade:s=!1}=t,{loop:r=!1}=t,{loopRandom:c=!1}=t,{scramble:i=!1}=t,{cursor:d=!0}=t,{delay:p=0}=t;const m=f(),u={interval:a,cascade:s,loop:r,loopRandom:c,scramble:i,cursor:d,delay:p,dispatch:m};return e.$$set=e=>{"interval"in e&&o(2,a=e.interval),"cascade"in e&&o(3,s=e.cascade),"loop"in e&&o(4,r=e.loop),"loopRandom"in e&&o(5,c=e.loopRandom),"scramble"in e&&o(6,i=e.scramble),"cursor"in e&&o(0,d=e.cursor),"delay"in e&&o(7,p=e.delay),"$$scope"in e&&o(8,l=e.$$scope)},[d,u,a,s,r,c,i,p,l,n]}class R extends e{constructor(e){var a;super(),document.getElementById("svelte-8gtb9p-style")||((a=n("style")).id="svelte-8gtb9p-style",a.textContent="@keyframes svelte-8gtb9p-cursorFade{0%,100%{opacity:1}50%{opacity:0}}.cursor.svelte-8gtb9p .typing::after{content:'▌';display:inline-block;color:var(--cursor-color);animation:svelte-8gtb9p-cursorFade 1.25s infinite}.delay.svelte-8gtb9p{visibility:hidden}",l(document.head,a)),t(this,e,M,E,o,{interval:2,cascade:3,loop:4,loopRandom:5,scramble:6,cursor:0,delay:7})}}export{R as T,w as a,x as c,g as h,L as l,b as r,v as s,h as t};
+import{SvelteComponent as e,init as t,safe_not_equal as o,element as l,append as n,create_slot as a,set_style as r,attr as s,toggle_class as c,insert as i,action_destroyer as d,update_slot as p,transition_in as m,transition_out as u,detach as y}from"svelte/internal";function b(){const e=function(){throw new Error("Function called outside component initialization")}();return(t,o)=>{e.$$.callbacks[t]}}const v=(e,t)=>{const o=document.createElement(t);return o.textContent=e,o},f=e=>1===e.childNodes.length&&3===e.childNodes[0].nodeType,g=e=>{if(f(e)){const t=e.textContent,o=v(e.textContent,"p");return e.textContent="",e.appendChild(o),[{currentNode:o,text:t}]}return[...e.children].map((e=>{const t=e.innerHTML.replaceAll("&amp;","&");return{currentNode:e,text:t}}))},$=async(e,t)=>{const{default:o}=t.loop&&await import("./loop.mjs")||t.loopRandom&&await import("./loopRandom.mjs")||t.cascade&&await import("./cascade.mjs")||t.scramble&&await import("./scramble.mjs")||await import("./default.mjs"),l=g(e);if(t.delay>0){const{sleep:o}=await import("./index2.mjs");await o(t.delay),e.classList.remove("delay")}o({node:e,elements:l},t)};function h(e){let t,o,n,b,v;const f=e[9].default,g=a(f,e,e[8],null);return{c(){t=l("div"),g&&g.c(),r(t,"--cursor-color","string"==typeof e[0]?e[0]:"black"),s(t,"class","svelte-8gtb9p"),c(t,"cursor",e[0]),c(t,"delay",e[1].delay>0)},m(l,a){i(l,t,a),g&&g.m(t,null),n=!0,b||(v=d(o=$.call(null,t,e[1])),b=!0)},p(e,[o]){g&&g.p&&256&o&&p(g,f,e,e[8],o,null,null),(!n||1&o)&&r(t,"--cursor-color","string"==typeof e[0]?e[0]:"black"),1&o&&c(t,"cursor",e[0]),2&o&&c(t,"delay",e[1].delay>0)},i(e){n||(m(g,e),n=!0)},o(e){u(g,e),n=!1},d(e){e&&y(t),g&&g.d(e),b=!1,v()}}}function x(e,t,o){let{$$slots:l={},$$scope:n}=t,{interval:a=30}=t,{cascade:r=!1}=t,{loop:s=!1}=t,{loopRandom:c=!1}=t,{scramble:i=!1}=t,{cursor:d=!0}=t,{delay:p=0}=t;const m=b(),u={interval:a,cascade:r,loop:s,loopRandom:c,scramble:i,cursor:d,delay:p,dispatch:m};return e.$$set=e=>{"interval"in e&&o(2,a=e.interval),"cascade"in e&&o(3,r=e.cascade),"loop"in e&&o(4,s=e.loop),"loopRandom"in e&&o(5,c=e.loopRandom),"scramble"in e&&o(6,i=e.scramble),"cursor"in e&&o(0,d=e.cursor),"delay"in e&&o(7,p=e.delay),"$$scope"in e&&o(8,n=e.$$scope)},[d,u,a,r,s,c,i,p,n,l]}class w extends e{constructor(e){var a;super(),document.getElementById("svelte-8gtb9p-style")||((a=l("style")).id="svelte-8gtb9p-style",a.textContent="@keyframes svelte-8gtb9p-cursorFade{0%,100%{opacity:1}50%{opacity:0}}.cursor.svelte-8gtb9p .typing::after{content:'▌';display:inline-block;color:var(--cursor-color);animation:svelte-8gtb9p-cursorFade 1.25s infinite}.delay.svelte-8gtb9p{visibility:hidden}",n(document.head,a)),t(this,e,x,h,o,{interval:2,cascade:3,loop:4,loopRandom:5,scramble:6,cursor:0,delay:7})}}export{w as T,v as c,g,f as h};
