+++ lib/esm/cascade.mjs
@@ -1,1 +1,1 @@
-import"svelte/internal";import{t}from"./Typewriter.mjs";const e=async({currentNode:e,text:n},s)=>{e.classList.add("typing");for(let i=0;i<=n.length;i++){"<"===n[i]&&(i=n.indexOf(">",i)),e.innerHTML=n.slice(0,i),await t(s.interval)}null!==e.nextSibling&&1==e.classList.length?e.removeAttribute("class"):e.classList.remove("typing")};export default async({elements:t},n)=>{t.forEach((({currentNode:t})=>t.textContent=""));for(const s of t)await e(s,n);n.dispatch("done")};
+import"./sleep.mjs";import{t}from"./typingInterval.mjs";const e=async({currentNode:e,text:n},s)=>{e.classList.add("typing");for(let i=0;i<=n.length;i++){"<"===n[i]&&(i=n.indexOf(">",i)),e.innerHTML=n.slice(0,i),await t(s.interval)}null!==e.nextSibling&&1==e.classList.length?e.removeAttribute("class"):e.classList.remove("typing")};export default async({elements:t},n)=>{t.forEach((({currentNode:t})=>t.textContent=""));for(const s of t)await e(s,n);n.dispatch("done")};
