+++ lib/esm/cleanChildNodes.mjs
@@ -1,0 +1,1 @@
+const o=o=>o.childNodes.forEach((o=>o.remove()));export{o as c};
