+++ lib/esm/loop.mjs
@@ -1,1 +1,1 @@
-import"svelte/internal";import{c as e,l as t}from"./Typewriter.mjs";const o=e=>e.childNodes.forEach((e=>e.remove()));export default async({node:r,elements:a},s)=>{for(;s.loop;)for(const{currentNode:n,text:c}of a){o(r);const a=n.tagName.toLowerCase(),i=[...n.attributes],l=e(c,a);i.forEach((({name:e,value:t})=>l.setAttribute(e,t))),r.appendChild(l),await t({currentNode:l,text:c},s),o(r)}};
+import"svelte/internal";import{c as e}from"./Typewriter.mjs";import"./sleep.mjs";import"./typingInterval.mjs";import{l as t}from"./loopTypewriterEffect.mjs";const o=e=>e.childNodes.forEach((e=>e.remove()));export default async({node:r,elements:a},s)=>{for(;s.loop;)for(const{currentNode:m,text:i}of a){o(r);const a=m.tagName.toLowerCase(),n=[...m.attributes],p=e(i,a);n.forEach((({name:e,value:t})=>p.setAttribute(e,t))),r.appendChild(p),await t({currentNode:p,text:i},s),o(r)}};
