+++ lib/esm/loopRandom.mjs
@@ -1,1 +1,1 @@
-import"svelte/internal";import{a as e,c as t,l as r,r as a,h as o}from"./Typewriter.mjs";let n=[];const s=e=>{for(;;){const t=a(0,e.length),r="number"==typeof n&&t!==n;if(Array.isArray(n)&&!n.includes(t)||r){r&&(n=[]),n.push(t);return e[t]}n.length===e.length&&(n=o?[]:n.pop())}};export default async({node:a,elements:o},n)=>{for(;n.loopRandom;){const{currentNode:l,text:i}=s(o);e(a);const p=l.tagName.toLowerCase(),c=t(i,p);[...l.attributes].forEach((({name:e,value:t})=>c.setAttribute(e,t))),a.appendChild(c),await r({currentNode:c,text:i},n),e(a)}};
+import"svelte/internal";import{c as e,h as t}from"./Typewriter.mjs";import{c as r}from"./cleanChildNodes.mjs";import"./sleep.mjs";import{r as o}from"./typingInterval.mjs";import{l as s}from"./loopTypewriterEffect.mjs";let a=[];const n=e=>{for(;;){const r=o(0,e.length),s="number"==typeof a&&r!==a;if(Array.isArray(a)&&!a.includes(r)||s){s&&(a=[]),a.push(r);return e[r]}a.length===e.length&&(a=t?[]:a.pop())}};export default async({node:t,elements:o},a)=>{for(;a.loopRandom;){const{currentNode:m,text:p}=n(o);r(t);const i=m.tagName.toLowerCase(),l=e(p,i);[...m.attributes].forEach((({name:e,value:t})=>l.setAttribute(e,t))),t.appendChild(l),await s({currentNode:l,text:p},a),r(t)}};
