+++ lib/esm/loopTypewriterEffect.mjs
@@ -1,0 +1,1 @@
+import{s as t}from"./sleep.mjs";import{t as e}from"./typingInterval.mjs";const n=async(t,n)=>{const a=t.innerHTML.replaceAll("&amp;","&");for(let l=a.length-1;l>=0;l--){">"===a[l]&&(l=a.lastIndexOf("<",l)),t.innerHTML=a.slice(0,l),await e(n.interval)}},a=async({currentNode:a,text:l},s)=>{a.textContent="",a.classList.add("typing");for(let i=0;i<=l.length;i++){"<"===l[i]&&(i=l.indexOf(">",i)+1),a.innerHTML=l.slice(0,i);const r=l.replaceAll("&","&amp;");a.innerHTML===r&&(s.dispatch("done"),await t("number"==typeof s.loop?s.loop:1500),await n(a,s)),await e(s.interval)}null!==a.nextSibling&&1==a.classList.length?a.removeAttribute("class"):a.classList.remove("typing")};export{a as l,n as u};
