+++ lib/esm/sleep.mjs
@@ -1,0 +1,1 @@
+const e=e=>new Promise((s=>setTimeout(s,e)));export{e as s};
