+++ lib/esm/typingInterval.mjs
@@ -1,0 +1,1 @@
+import{s}from"./sleep.mjs";const a=(s,a)=>Math.floor(Math.random()*(a-s)+s),o=async o=>s(o[a(0,o.length)]||o);export{a as r,o as t};
