+++ lib/umd/index.js
@@ -1,1 +1,1 @@
-!function(t,e){"object"==typeof exports&&"undefined"!=typeof module?module.exports=e():"function"==typeof define&&define.amd?define(e):(t="undefined"!=typeof globalThis?globalThis:t||self).Typewriter=e()}(this,(function(){"use strict";function t(){}function e(t){return t()}function n(){return Object.create(null)}function o(t){t.forEach(e)}function r(t){return"function"==typeof t}function s(t,e){return t!=t?e==e:t!==e||t&&"object"==typeof t||"function"==typeof t}function c(t,e,n,o){return t[1]&&o?function(t,e){for(const n in e)t[n]=e[n];return t}(n.ctx.slice(),t[1](o(e))):n.ctx}function a(t,e,n,o,r,s,a){const l=function(t,e,n,o){if(t[2]&&o){const r=t[2](o(n));if(void 0===e.dirty)return r;if("object"==typeof r){const t=[],n=Math.max(e.dirty.length,r.length);for(let o=0;o<n;o+=1)t[o]=e.dirty[o]|r[o];return t}return e.dirty|r}return e.dirty}(e,o,r,s);if(l){const r=c(e,n,o,a);t.p(r,l)}}function l(t){t.parentNode.removeChild(t)}function i(t){return document.createElement(t)}function u(t,e,n,o){t.style.setProperty(e,n,o?"important":"")}function d(t,e,n){t.classList[n?"add":"remove"](e)}let f;function p(t){f=t}function m(){const t=function(){if(!f)throw new Error("Function called outside component initialization");return f}();return(e,n)=>{const o=t.$$.callbacks[e];if(o){const r=function(t,e){const n=document.createEvent("CustomEvent");return n.initCustomEvent(t,!1,!1,e),n}(e,n);o.slice().forEach((e=>{e.call(t,r)}))}}}const h=[],y=[],g=[],b=[],v=Promise.resolve();let $=!1;function x(t){g.push(t)}let _=!1;const w=new Set;function L(){if(!_){_=!0;do{for(let t=0;t<h.length;t+=1){const e=h[t];p(e),N(e.$$)}for(p(null),h.length=0;y.length;)y.pop()();for(let t=0;t<g.length;t+=1){const e=g[t];w.has(e)||(w.add(e),e())}g.length=0}while(h.length);for(;b.length;)b.pop()();$=!1,_=!1,w.clear()}}function N(t){if(null!==t.fragment){t.update(),o(t.before_update);const e=t.dirty;t.dirty=[-1],t.fragment&&t.fragment.p(t.ctx,e),t.after_update.forEach(x)}}const E=new Set;function T(t,e){t&&t.i&&(E.delete(t),t.i(e))}function M(t,e){-1===t.$$.dirty[0]&&(h.push(t),$||($=!0,v.then(L)),t.$$.dirty.fill(0)),t.$$.dirty[e/31|0]|=1<<e%31}function C(s,c,a,i,u,d,m=[-1]){const h=f;p(s);const y=c.props||{},g=s.$$={fragment:null,ctx:null,props:d,update:t,not_equal:u,bound:n(),on_mount:[],on_destroy:[],before_update:[],after_update:[],context:new Map(h?h.$$.context:[]),callbacks:n(),dirty:m,skip_bound:!1};let b=!1;if(g.ctx=a?a(s,y,((t,e,...n)=>{const o=n.length?n[0]:e;return g.ctx&&u(g.ctx[t],g.ctx[t]=o)&&(!g.skip_bound&&g.bound[t]&&g.bound[t](o),b&&M(s,t)),e})):[],g.update(),b=!0,o(g.before_update),g.fragment=!!i&&i(g.ctx),c.target){if(c.hydrate){const t=function(t){return Array.from(t.childNodes)}(c.target);g.fragment&&g.fragment.l(t),t.forEach(l)}else g.fragment&&g.fragment.c();c.intro&&T(s.$$.fragment),function(t,n,s){const{fragment:c,on_mount:a,on_destroy:l,after_update:i}=t.$$;c&&c.m(n,s),x((()=>{const n=a.map(e).filter(r);l?l.push(...n):o(n),t.$$.on_mount=[]})),i.forEach(x)}(s,c.target,c.anchor),L()}p(h)}const k=t=>new Promise((e=>setTimeout(e,t))),A=(t,e)=>Math.floor(Math.random()*(e-t)+t),O=t=>1===t.childNodes.length&&3===t.childNodes[0].nodeType,H=async t=>k(t[A(0,t.length)]||t),j=t=>t.childNodes.forEach((t=>t.remove())),P=(t,e)=>{const n=document.createElement(e);return n.textContent=t,n},z=t=>{if(O(t)){const e=t.textContent,n=P(t.textContent,"p");return t.textContent="",t.appendChild(n),[{currentNode:n,text:e}]}return[...t.children].map((t=>{const e=t.innerHTML.replaceAll("&amp;","&");return{currentNode:t,text:e}}))},R=async(t,e)=>{const n=t.innerHTML.replaceAll("&amp;","&");for(let o=n.length-1;o>=0;o--){">"===n[o]&&(o=n.lastIndexOf("<",o)),t.innerHTML=n.slice(0,o),await H(e.interval)}},S=async({currentNode:t,text:e},n)=>{t.textContent="",t.classList.add("typing");for(let o=0;o<=e.length;o++){"<"===e[o]&&(o=e.indexOf(">",o)+1),t.innerHTML=e.slice(0,o);const r=e.replaceAll("&","&amp;");t.innerHTML===r&&(n.dispatch("done"),await k("number"==typeof n.loop?n.loop:1500),await R(t,n)),await H(n.interval)}null!==t.nextSibling&&1==t.classList.length?t.removeAttribute("class"):t.classList.remove("typing")};var I=Object.freeze({__proto__:null,sleep:k,rng:A,hasSingleTextNode:O,typingInterval:H,cleanChildNodes:j,createElement:P,getElements:z,loopTypewriterEffect:S});const F=async(t,e)=>{const{default:n}=e.loop&&await Promise.resolve().then((function(){return G}))||e.loopRandom&&await Promise.resolve().then((function(){return Q}))||e.cascade&&await Promise.resolve().then((function(){return V}))||e.scramble&&await Promise.resolve().then((function(){return tt}))||await Promise.resolve().then((function(){return nt})),o=z(t);if(e.delay>0){const{sleep:n}=await Promise.resolve().then((function(){return I}));await n(e.delay),t.classList.remove("delay")}n({node:t,elements:o},e)};function B(e){let n,o,s,f,p;const m=e[9].default,h=function(t,e,n,o){if(t){const r=c(t,e,n,o);return t[0](r)}}(m,e,e[8],null);return{c(){var t,o,r;n=i("div"),h&&h.c(),u(n,"--cursor-color","string"==typeof e[0]?e[0]:"black"),t=n,o="class",null==(r="svelte-8gtb9p")?t.removeAttribute(o):t.getAttribute(o)!==r&&t.setAttribute(o,r),d(n,"cursor",e[0]),d(n,"delay",e[1].delay>0)},m(c,a){var l;!function(t,e,n){t.insertBefore(e,n||null)}(c,n,a),h&&h.m(n,null),s=!0,f||(l=o=F.call(null,n,e[1]),p=l&&r(l.destroy)?l.destroy:t,f=!0)},p(t,[e]){h&&h.p&&256&e&&a(h,m,t,t[8],e,null,null),(!s||1&e)&&u(n,"--cursor-color","string"==typeof t[0]?t[0]:"black"),1&e&&d(n,"cursor",t[0]),2&e&&d(n,"delay",t[1].delay>0)},i(t){s||(T(h,t),s=!0)},o(t){!function(t,e,n,o){if(t&&t.o){if(E.has(t))return;E.add(t),(void 0).c.push((()=>{E.delete(t),o&&(n&&t.d(1),o())})),t.o(e)}}(h,t),s=!1},d(t){t&&l(n),h&&h.d(t),f=!1,p()}}}function D(t,e,n){let{$$slots:o={},$$scope:r}=e,{interval:s=30}=e,{cascade:c=!1}=e,{loop:a=!1}=e,{loopRandom:l=!1}=e,{scramble:i=!1}=e,{cursor:u=!0}=e,{delay:d=0}=e;const f=m(),p={interval:s,cascade:c,loop:a,loopRandom:l,scramble:i,cursor:u,delay:d,dispatch:f};return t.$$set=t=>{"interval"in t&&n(2,s=t.interval),"cascade"in t&&n(3,c=t.cascade),"loop"in t&&n(4,a=t.loop),"loopRandom"in t&&n(5,l=t.loopRandom),"scramble"in t&&n(6,i=t.scramble),"cursor"in t&&n(0,u=t.cursor),"delay"in t&&n(7,d=t.delay),"$$scope"in t&&n(8,r=t.$$scope)},[u,p,s,c,a,l,i,d,r,o]}const q=t=>t.childNodes.forEach((t=>t.remove()));var G=Object.freeze({__proto__:null,default:async({node:t,elements:e},n)=>{for(;n.loop;)for(const{currentNode:o,text:r}of e){q(t);const e=o.tagName.toLowerCase(),s=[...o.attributes],c=P(r,e);s.forEach((({name:t,value:e})=>c.setAttribute(t,e))),t.appendChild(c),await S({currentNode:c,text:r},n),q(t)}}});let J=[];const K=t=>{for(;;){const e=A(0,t.length),n="number"==typeof J&&e!==J;if(Array.isArray(J)&&!J.includes(e)||n){n&&(J=[]),J.push(e);return t[e]}J.length===t.length&&(J=O?[]:J.pop())}};var Q=Object.freeze({__proto__:null,default:async({node:t,elements:e},n)=>{for(;n.loopRandom;){const{currentNode:o,text:r}=K(e);j(t);const s=o.tagName.toLowerCase(),c=P(r,s);[...o.attributes].forEach((({name:t,value:e})=>c.setAttribute(t,e))),t.appendChild(c),await S({currentNode:c,text:r},n),j(t)}}});const U=async({currentNode:t,text:e},n)=>{t.classList.add("typing");for(let o=0;o<=e.length;o++){"<"===e[o]&&(o=e.indexOf(">",o)),t.innerHTML=e.slice(0,o),await H(n.interval)}null!==t.nextSibling&&1==t.classList.length?t.removeAttribute("class"):t.classList.remove("typing")};var V=Object.freeze({__proto__:null,default:async({elements:t},e)=>{t.forEach((({currentNode:t})=>t.textContent=""));for(const n of t)await U(n,e);e.dispatch("done")}});let W=[];const X=t=>W.find((e=>e.currentNode===t)),Y=t=>{const e=(t=>{const e=/(<([^>]+)>)/g,n=[];let o=[];for(;null!==(o=e.exec(t.innerHTML));){const t=o.index,r=e.lastIndex;n.push([t,r-1])}return n})(t),n=t.innerHTML.split("").map(((n,o)=>{const{matchingLetters:r}=X(t),s=r.includes(o)||n.match(/\s+/g)||((t,e)=>e.some((([e,n])=>{var o;return(o=t)>=e&&o<=n})))(o,e),c=String.fromCharCode(65+Math.round(50*Math.random()));return s?n:c})).join("");t.innerHTML=n},Z=(t,e)=>{const n=t.innerHTML;for(let o=0;o<e.length;o++){const r=e[o],{matchingLetters:s}=X(t);!s.includes(o)&&r===n[o]&&s.push(o)}};var tt=Object.freeze({__proto__:null,default:async({elements:t},e)=>{W=[...t.map((({currentNode:t})=>({currentNode:t,matchingLetters:[]})))],await new Promise((n=>{t.forEach((async({currentNode:t,text:o})=>{const r="number"==typeof e.scramble?e.scramble:3e3,s=(new Date).getTime();for(;;){Y(t),Z(t,o),await k(e.interval);const c=t.innerHTML!=o,a=(new Date).getTime()-s<r;if(!c||!a){n();break}}t.innerHTML=o}))})),e.dispatch("done")}});const et=async({currentNode:t,text:e},n)=>{t.classList.add("typing");for(let o=0;o<=e.length;o++){"<"===e[o]&&(o=e.indexOf(">",o)),t.innerHTML=e.slice(0,o),await H(n.interval)}null!==t.nextSibling&&1==t.classList.length?t.removeAttribute("class"):t.classList.remove("typing")};var nt=Object.freeze({__proto__:null,default:async({elements:t},e)=>{for(const n of t)et(n,e);const{currentNode:n}=t.reduce(((t,e)=>{const n=t.text.length;return e.text.length>n?t=e:t}));new MutationObserver((t=>{t.forEach((t=>{const n=!t.target.classList.contains("typing");"attributes"===t.type&&n&&e.dispatch("done")}))})).observe(n,{attributes:!0,childList:!0,subtree:!0})}});return class extends class{$destroy(){!function(t,e){const n=t.$$;null!==n.fragment&&(o(n.on_destroy),n.fragment&&n.fragment.d(e),n.on_destroy=n.fragment=null,n.ctx=[])}(this,1),this.$destroy=t}$on(t,e){const n=this.$$.callbacks[t]||(this.$$.callbacks[t]=[]);return n.push(e),()=>{const t=n.indexOf(e);-1!==t&&n.splice(t,1)}}$set(t){var e;this.$$set&&(e=t,0!==Object.keys(e).length)&&(this.$$.skip_bound=!0,this.$$set(t),this.$$.skip_bound=!1)}}{constructor(t){var e,n,o;super(),document.getElementById("svelte-8gtb9p-style")||((o=i("style")).id="svelte-8gtb9p-style",o.textContent="@keyframes svelte-8gtb9p-cursorFade{0%,100%{opacity:1}50%{opacity:0}}.cursor.svelte-8gtb9p .typing::after{content:'▌';display:inline-block;color:var(--cursor-color);animation:svelte-8gtb9p-cursorFade 1.25s infinite}.delay.svelte-8gtb9p{visibility:hidden}",e=document.head,n=o,e.appendChild(n)),C(this,t,D,B,s,{interval:2,cascade:3,loop:4,loopRandom:5,scramble:6,cursor:0,delay:7})}}}));
+!function(t,e){"object"==typeof exports&&"undefined"!=typeof module?module.exports=e():"function"==typeof define&&define.amd?define(e):(t="undefined"!=typeof globalThis?globalThis:t||self).Typewriter=e()}(this,(function(){"use strict";function t(){}function e(t){return t()}function n(){return Object.create(null)}function o(t){t.forEach(e)}function r(t){return"function"==typeof t}function s(t,e){return t!=t?e==e:t!==e||t&&"object"==typeof t||"function"==typeof t}function c(t,e,n,o){return t[1]&&o?function(t,e){for(const n in e)t[n]=e[n];return t}(n.ctx.slice(),t[1](o(e))):n.ctx}function a(t,e,n,o,r,s,a){const l=function(t,e,n,o){if(t[2]&&o){const r=t[2](o(n));if(void 0===e.dirty)return r;if("object"==typeof r){const t=[],n=Math.max(e.dirty.length,r.length);for(let o=0;o<n;o+=1)t[o]=e.dirty[o]|r[o];return t}return e.dirty|r}return e.dirty}(e,o,r,s);if(l){const r=c(e,n,o,a);t.p(r,l)}}function l(t){t.parentNode.removeChild(t)}function i(t){return document.createElement(t)}function u(t,e,n,o){t.style.setProperty(e,n,o?"important":"")}function d(t,e,n){t.classList[n?"add":"remove"](e)}let f;function p(t){f=t}function m(){const t=function(){if(!f)throw new Error("Function called outside component initialization");return f}();return(e,n)=>{const o=t.$$.callbacks[e];if(o){const r=function(t,e){const n=document.createEvent("CustomEvent");return n.initCustomEvent(t,!1,!1,e),n}(e,n);o.slice().forEach((e=>{e.call(t,r)}))}}}const h=[],y=[],g=[],b=[],v=Promise.resolve();let $=!1;function x(t){g.push(t)}let _=!1;const w=new Set;function L(){if(!_){_=!0;do{for(let t=0;t<h.length;t+=1){const e=h[t];p(e),N(e.$$)}for(p(null),h.length=0;y.length;)y.pop()();for(let t=0;t<g.length;t+=1){const e=g[t];w.has(e)||(w.add(e),e())}g.length=0}while(h.length);for(;b.length;)b.pop()();$=!1,_=!1,w.clear()}}function N(t){if(null!==t.fragment){t.update(),o(t.before_update);const e=t.dirty;t.dirty=[-1],t.fragment&&t.fragment.p(t.ctx,e),t.after_update.forEach(x)}}const E=new Set;function T(t,e){t&&t.i&&(E.delete(t),t.i(e))}function M(t,e){-1===t.$$.dirty[0]&&(h.push(t),$||($=!0,v.then(L)),t.$$.dirty.fill(0)),t.$$.dirty[e/31|0]|=1<<e%31}function C(s,c,a,i,u,d,m=[-1]){const h=f;p(s);const y=c.props||{},g=s.$$={fragment:null,ctx:null,props:d,update:t,not_equal:u,bound:n(),on_mount:[],on_destroy:[],before_update:[],after_update:[],context:new Map(h?h.$$.context:[]),callbacks:n(),dirty:m,skip_bound:!1};let b=!1;if(g.ctx=a?a(s,y,((t,e,...n)=>{const o=n.length?n[0]:e;return g.ctx&&u(g.ctx[t],g.ctx[t]=o)&&(!g.skip_bound&&g.bound[t]&&g.bound[t](o),b&&M(s,t)),e})):[],g.update(),b=!0,o(g.before_update),g.fragment=!!i&&i(g.ctx),c.target){if(c.hydrate){const t=function(t){return Array.from(t.childNodes)}(c.target);g.fragment&&g.fragment.l(t),t.forEach(l)}else g.fragment&&g.fragment.c();c.intro&&T(s.$$.fragment),function(t,n,s){const{fragment:c,on_mount:a,on_destroy:l,after_update:i}=t.$$;c&&c.m(n,s),x((()=>{const n=a.map(e).filter(r);l?l.push(...n):o(n),t.$$.on_mount=[]})),i.forEach(x)}(s,c.target,c.anchor),L()}p(h)}const k=t=>t.childNodes.forEach((t=>t.remove())),A=(t,e)=>{const n=document.createElement(e);return n.textContent=t,n},O=t=>1===t.childNodes.length&&3===t.childNodes[0].nodeType,H=t=>{if(O(t)){const e=t.textContent,n=A(t.textContent,"p");return t.textContent="",t.appendChild(n),[{currentNode:n,text:e}]}return[...t.children].map((t=>{const e=t.innerHTML.replaceAll("&amp;","&");return{currentNode:t,text:e}}))},j=t=>new Promise((e=>setTimeout(e,t))),P=(t,e)=>Math.floor(Math.random()*(e-t)+t),z=async t=>j(t[P(0,t.length)]||t),R=async(t,e)=>{const n=t.innerHTML.replaceAll("&amp;","&");for(let o=n.length-1;o>=0;o--){">"===n[o]&&(o=n.lastIndexOf("<",o)),t.innerHTML=n.slice(0,o),await z(e.interval)}},S=async({currentNode:t,text:e},n)=>{t.textContent="",t.classList.add("typing");for(let o=0;o<=e.length;o++){"<"===e[o]&&(o=e.indexOf(">",o)+1),t.innerHTML=e.slice(0,o);const r=e.replaceAll("&","&amp;");t.innerHTML===r&&(n.dispatch("done"),await j("number"==typeof n.loop?n.loop:1500),await R(t,n)),await z(n.interval)}null!==t.nextSibling&&1==t.classList.length?t.removeAttribute("class"):t.classList.remove("typing")};var I=Object.freeze({__proto__:null,cleanChildNodes:k,createElement:A,getElements:H,hasSingleTextNode:O,loopTypewriterEffect:S,rng:P,sleep:j,typingInterval:z,unwriteEffect:R});const F=async(t,e)=>{const{default:n}=e.loop&&await Promise.resolve().then((function(){return G}))||e.loopRandom&&await Promise.resolve().then((function(){return Q}))||e.cascade&&await Promise.resolve().then((function(){return V}))||e.scramble&&await Promise.resolve().then((function(){return tt}))||await Promise.resolve().then((function(){return nt})),o=H(t);if(e.delay>0){const{sleep:n}=await Promise.resolve().then((function(){return I}));await n(e.delay),t.classList.remove("delay")}n({node:t,elements:o},e)};function B(e){let n,o,s,f,p;const m=e[9].default,h=function(t,e,n,o){if(t){const r=c(t,e,n,o);return t[0](r)}}(m,e,e[8],null);return{c(){var t,o,r;n=i("div"),h&&h.c(),u(n,"--cursor-color","string"==typeof e[0]?e[0]:"black"),t=n,o="class",null==(r="svelte-8gtb9p")?t.removeAttribute(o):t.getAttribute(o)!==r&&t.setAttribute(o,r),d(n,"cursor",e[0]),d(n,"delay",e[1].delay>0)},m(c,a){var l;!function(t,e,n){t.insertBefore(e,n||null)}(c,n,a),h&&h.m(n,null),s=!0,f||(l=o=F.call(null,n,e[1]),p=l&&r(l.destroy)?l.destroy:t,f=!0)},p(t,[e]){h&&h.p&&256&e&&a(h,m,t,t[8],e,null,null),(!s||1&e)&&u(n,"--cursor-color","string"==typeof t[0]?t[0]:"black"),1&e&&d(n,"cursor",t[0]),2&e&&d(n,"delay",t[1].delay>0)},i(t){s||(T(h,t),s=!0)},o(t){!function(t,e,n,o){if(t&&t.o){if(E.has(t))return;E.add(t),(void 0).c.push((()=>{E.delete(t),o&&(n&&t.d(1),o())})),t.o(e)}}(h,t),s=!1},d(t){t&&l(n),h&&h.d(t),f=!1,p()}}}function D(t,e,n){let{$$slots:o={},$$scope:r}=e,{interval:s=30}=e,{cascade:c=!1}=e,{loop:a=!1}=e,{loopRandom:l=!1}=e,{scramble:i=!1}=e,{cursor:u=!0}=e,{delay:d=0}=e;const f=m(),p={interval:s,cascade:c,loop:a,loopRandom:l,scramble:i,cursor:u,delay:d,dispatch:f};return t.$$set=t=>{"interval"in t&&n(2,s=t.interval),"cascade"in t&&n(3,c=t.cascade),"loop"in t&&n(4,a=t.loop),"loopRandom"in t&&n(5,l=t.loopRandom),"scramble"in t&&n(6,i=t.scramble),"cursor"in t&&n(0,u=t.cursor),"delay"in t&&n(7,d=t.delay),"$$scope"in t&&n(8,r=t.$$scope)},[u,p,s,c,a,l,i,d,r,o]}const q=t=>t.childNodes.forEach((t=>t.remove()));var G=Object.freeze({__proto__:null,default:async({node:t,elements:e},n)=>{for(;n.loop;)for(const{currentNode:o,text:r}of e){q(t);const e=o.tagName.toLowerCase(),s=[...o.attributes],c=A(r,e);s.forEach((({name:t,value:e})=>c.setAttribute(t,e))),t.appendChild(c),await S({currentNode:c,text:r},n),q(t)}}});let J=[];const K=t=>{for(;;){const e=P(0,t.length),n="number"==typeof J&&e!==J;if(Array.isArray(J)&&!J.includes(e)||n){n&&(J=[]),J.push(e);return t[e]}J.length===t.length&&(J=O?[]:J.pop())}};var Q=Object.freeze({__proto__:null,default:async({node:t,elements:e},n)=>{for(;n.loopRandom;){const{currentNode:o,text:r}=K(e);k(t);const s=o.tagName.toLowerCase(),c=A(r,s);[...o.attributes].forEach((({name:t,value:e})=>c.setAttribute(t,e))),t.appendChild(c),await S({currentNode:c,text:r},n),k(t)}}});const U=async({currentNode:t,text:e},n)=>{t.classList.add("typing");for(let o=0;o<=e.length;o++){"<"===e[o]&&(o=e.indexOf(">",o)),t.innerHTML=e.slice(0,o),await z(n.interval)}null!==t.nextSibling&&1==t.classList.length?t.removeAttribute("class"):t.classList.remove("typing")};var V=Object.freeze({__proto__:null,default:async({elements:t},e)=>{t.forEach((({currentNode:t})=>t.textContent=""));for(const n of t)await U(n,e);e.dispatch("done")}});let W=[];const X=t=>W.find((e=>e.currentNode===t)),Y=t=>{const e=(t=>{const e=/(<([^>]+)>)/g,n=[];let o=[];for(;null!==(o=e.exec(t.innerHTML));){const t=o.index,r=e.lastIndex;n.push([t,r-1])}return n})(t),n=t.innerHTML.split("").map(((n,o)=>{const{matchingLetters:r}=X(t),s=r.includes(o)||n.match(/\s+/g)||((t,e)=>e.some((([e,n])=>{var o;return(o=t)>=e&&o<=n})))(o,e),c=String.fromCharCode(65+Math.round(50*Math.random()));return s?n:c})).join("");t.innerHTML=n},Z=(t,e)=>{const n=t.innerHTML;for(let o=0;o<e.length;o++){const r=e[o],{matchingLetters:s}=X(t);!s.includes(o)&&r===n[o]&&s.push(o)}};var tt=Object.freeze({__proto__:null,default:async({elements:t},e)=>{W=[...t.map((({currentNode:t})=>({currentNode:t,matchingLetters:[]})))],await new Promise((n=>{t.forEach((async({currentNode:t,text:o})=>{const r="number"==typeof e.scramble?e.scramble:3e3,s=(new Date).getTime();for(;;){Y(t),Z(t,o),await j(e.interval);const c=t.innerHTML!=o,a=(new Date).getTime()-s<r;if(!c||!a){n();break}}t.innerHTML=o}))})),e.dispatch("done")}});const et=async({currentNode:t,text:e},n)=>{t.classList.add("typing");for(let o=0;o<=e.length;o++){"<"===e[o]&&(o=e.indexOf(">",o)),t.innerHTML=e.slice(0,o),await z(n.interval)}null!==t.nextSibling&&1==t.classList.length?t.removeAttribute("class"):t.classList.remove("typing")};var nt=Object.freeze({__proto__:null,default:async({elements:t},e)=>{for(const n of t)et(n,e);const{currentNode:n}=t.reduce(((t,e)=>{const n=t.text.length;return e.text.length>n?t=e:t}));new MutationObserver((t=>{t.forEach((t=>{const n=!t.target.classList.contains("typing");"attributes"===t.type&&n&&e.dispatch("done")}))})).observe(n,{attributes:!0,childList:!0,subtree:!0})}});return class extends class{$destroy(){!function(t,e){const n=t.$$;null!==n.fragment&&(o(n.on_destroy),n.fragment&&n.fragment.d(e),n.on_destroy=n.fragment=null,n.ctx=[])}(this,1),this.$destroy=t}$on(t,e){const n=this.$$.callbacks[t]||(this.$$.callbacks[t]=[]);return n.push(e),()=>{const t=n.indexOf(e);-1!==t&&n.splice(t,1)}}$set(t){var e;this.$$set&&(e=t,0!==Object.keys(e).length)&&(this.$$.skip_bound=!0,this.$$set(t),this.$$.skip_bound=!1)}}{constructor(t){var e,n,o;super(),document.getElementById("svelte-8gtb9p-style")||((o=i("style")).id="svelte-8gtb9p-style",o.textContent="@keyframes svelte-8gtb9p-cursorFade{0%,100%{opacity:1}50%{opacity:0}}.cursor.svelte-8gtb9p .typing::after{content:'▌';display:inline-block;color:var(--cursor-color);animation:svelte-8gtb9p-cursorFade 1.25s infinite}.delay.svelte-8gtb9p{visibility:hidden}",e=document.head,n=o,e.appendChild(n)),C(this,t,D,B,s,{interval:2,cascade:3,loop:4,loopRandom:5,scramble:6,cursor:0,delay:7})}}}));
