+++ src/helpers/createElement.js
@@ -1,0 +1,16 @@
+/// <reference path='../types.js' />
+
+/**
+ * Create a HTML element
+ * @param {string} text The text to be inserted into the HTML element
+ * @param {string} elementTag The tag name of the newly created HTML element
+ * @example
+ * convertTextToElement(parent)
+ */
+const createElement = (text, elementTag) => {
+	const element = document.createElement(elementTag)
+	element.textContent = text
+	return element
+}
+
+export { createElement }
