+++ src/helpers/sleep.js
@@ -1,0 +1,14 @@
+/// <reference path='../types.js' />
+
+/**
+ * Waits for a given amount of time
+ * @param {number} ms The time in milliseconds to wait before resuming execution
+ * @example
+ *  const someFunction = () => {
+ *    // Waits for 1 second before resuming execution
+ *    await sleep(1000)
+ *  }
+ */
+const sleep = ms => new Promise(resolve => setTimeout(resolve, ms))
+
+export { sleep }
