+++ src/helpers/unwriteEffect.js
@@ -1,0 +1,15 @@
+/// <reference path='../types.js' />
+
+import { typingInterval } from './typingInterval'
+
+const unwriteEffect = async (currentNode, options) => {
+	const text = currentNode.innerHTML.replaceAll('&amp;', '&')
+	for (let index = text.length - 1; index >= 0; index--) {
+		const letter = text[index]
+		letter === '>' && (index = text.lastIndexOf('<', index))
+		currentNode.innerHTML = text.slice(0, index)
+		await typingInterval(options.interval)
+	}
+}
+
+export { unwriteEffect }
