+++ src/modes/cascade.js
@@ -1,5 +1,5 @@
 /// <reference path='../types.js' />
-import { typingInterval } from '../utils'
+import { typingInterval } from '../helpers'
 
 // Unify the function below with the one on cascade mode
 /** @type {TypewriterEffectFn} */
